Upgrade react-hot-loader to react-transform
It changes the way you need to run the application. For now you have to run:

```
node server.js
```

and

```
vtex watch yoursandbox
```

This will change when Toolbelt adapts to the new changes.

Also, `webpack.config.js` was simplified and reordered, the things that are more
customizable were moved to the top; things that dont change quite often were left
on the bottom on the file.